Create CreateMachineHandler
